requirements_nicegui.txt
docxtpl==0.16.7
Jinja2==3.1.3
nicegui==1.4.18
openpyxl==3.1.2
pandas==2.2.2
PyPDF2==3.0.1
python-docx==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
